Move logo_url to knockout observable
The old `computed` function wasn't actually working correctly, since
user_data wasn't in a Knockout observable (and since `computed` watches
these to know when to recompute, it never recomputed).

Instead, this adds a callback when the asynchronous user_data request
completes, which then updates the value of a new logo_url observable.